+++ src/CommerceMolliePlusPlugin.php
@@ -2,7 +2,7 @@
 
 namespace white\commerce\mollie\plus;
 
-use Craft;
+use craft\base\Element;
 use craft\base\Plugin;
 use craft\commerce\elements\Order;
 use craft\commerce\Plugin as CommercePlugin;
@@ -17,23 +17,27 @@ use yii\base\Event;
  */
 class CommerceMolliePlusPlugin extends Plugin
 {
-    public function init()
+    public function init(): void
     {
         parent::init();
 
-        Event::on(Gateways::class, Gateways::EVENT_REGISTER_GATEWAY_TYPES,  function(RegisterComponentTypesEvent $event) {
-            $event->types[] = Gateway::class;
-        });
+        Event::on(
+            Gateways::class,
+            Gateways::EVENT_REGISTER_GATEWAY_TYPES,
+            function(RegisterComponentTypesEvent $event): void {
+                $event->types[] = Gateway::class;
+            }
+        );
 
         $this->registerOrderEventListeners();
     }
 
-    protected function registerOrderEventListeners()
+    protected function registerOrderEventListeners(): void
     {
         Event::on(
             Order::class,
-            Order::EVENT_AFTER_SAVE,
-            function (ModelEvent $event) {
+            Element::EVENT_AFTER_SAVE,
+            function(ModelEvent $event): void {
                 /** @var Order $order */
                 $order = $event->sender;
                 if ($order->propagating || !$order->orderStatusId) {
@@ -43,12 +47,10 @@ class CommerceMolliePlusPlugin extends Plugin
                 // Automatic transaction capture based on order status configured in the gateway settings
                 foreach ($order->getTransactions() as $transaction) {
                     $gateway = $transaction->getGateway();
-                    if ($gateway instanceof Gateway && $transaction->canCapture()) {
-                        if (is_array($gateway->orderStatusToCapture) && in_array($order->getOrderStatus()->handle, $gateway->orderStatusToCapture)) {
-                            $child = CommercePlugin::getInstance()->getPayments()->captureTransaction($transaction);
-                            if ($child->status == TransactionRecord::STATUS_SUCCESS) {
-                                $child->order->updateOrderPaidInformation();
-                            }
+                    if ($gateway instanceof Gateway && $transaction->canCapture() && (is_array($gateway->orderStatusToCapture) && in_array($order->getOrderStatus()->handle, $gateway->orderStatusToCapture))) {
+                        $child = CommercePlugin::getInstance()->getPayments()->captureTransaction($transaction);
+                        if ($child->status == TransactionRecord::STATUS_SUCCESS) {
+                            $child->getOrder()->updateOrderPaidInformation();
                         }
                     }
                 }