+++ src/gateways/Gateway.php
@@ -3,48 +3,88 @@
 namespace white\commerce\mollie\plus\gateways;
 
 use Craft;
-use craft\commerce\base\Purchasable;
 use craft\commerce\base\RequestResponseInterface;
 use craft\commerce\elements\Order;
 use craft\commerce\helpers\Currency;
 use craft\commerce\models\payments\BasePaymentForm;
 use craft\commerce\models\Transaction;
-use Omnipay\Common\Message\AbstractRequest;
-use white\commerce\mollie\plus\events\CreatePaymentRequestEvent;
-use white\commerce\mollie\plus\models\RequestResponse;
 use craft\commerce\omnipay\base\OffsiteGateway;
 use craft\commerce\Plugin as Commerce;
 use craft\commerce\records\Transaction as TransactionRecord;
+use craft\helpers\App;
 use craft\web\Response;
 use craft\web\View;
-use white\commerce\mollie\plus\models\forms\MollieOffsitePaymentForm;
 use Omnipay\Common\AbstractGateway;
+use Omnipay\Common\CreditCard;
+use Omnipay\Common\Exception\InvalidRequestException;
+use Omnipay\Common\ItemBag;
+use Omnipay\Common\Message\AbstractResponse;
 use Omnipay\Common\Message\RequestInterface;
 use Omnipay\Common\Message\ResponseInterface;
 use Omnipay\Mollie\Gateway as OmnipayGateway;
+use Omnipay\Mollie\Item;
+use Omnipay\Mollie\Message\Response\FetchPaymentMethodsResponse;
+use white\commerce\mollie\plus\events\CreatePaymentRequestEvent;
+use white\commerce\mollie\plus\models\forms\MollieOffsitePaymentForm;
+use white\commerce\mollie\plus\models\RequestResponse;
 use yii\base\NotSupportedException;
 
+/**
+ *
+ * @property-write null|string $apiKey
+ * @property-read null|string $settingsHtml
+ */
 class Gateway extends OffsiteGateway
 {
     /**
-     * @var string
+     * @var string|null
      */
-    public $apiKey;
+    public ?string $_apiKey = null;
 
     /**
      * @var array|string[]
      */
-    public $orderStatusToCapture = [];
+    public array $orderStatusToCapture = [];
 
     /**
      * @inheritdoc
      */
-    public $sendCartInfo = true;
+    public bool $sendCartInfo = true;
 
     /**
      * Event to mutate the request payload
+     * @var string
      */
-    const EVENT_CREATE_PAYMENT_REQUEST = 'createPaymentRequestEvent';
+    public const EVENT_CREATE_PAYMENT_REQUEST = 'createPaymentRequestEvent';
+
+    /**
+     * @inheritdoc
+     */
+    public function getSettings(): array
+    {
+        $settings = parent::getSettings();
+        $settings['apiKey'] = $this->getApiKey(false);
+
+        return $settings;
+    }
+
+    /**
+     * @param bool $parse
+     * @return string|null
+     */
+    public function getApiKey(bool $parse = true): ?string
+    {
+        return $parse ? App::parseEnv($this->_apiKey) : $this->_apiKey;
+    }
+
+    /**
+     * @param string|null $apiKey
+     * @return void
+     */
+    public function setApiKey(?string $apiKey): void
+    {
+        $this->_apiKey = $apiKey;
+    }
 
     /**
      * @inheritdoc
@@ -65,7 +105,7 @@ class Gateway extends OffsiteGateway
     /**
      * @inheritdoc
      */
-    public function rules()
+    public function rules(): array
     {
         $rules = parent::rules();
         $rules[] = ['paymentType', 'compare', 'compareValue' => 'authorize'];
@@ -86,7 +126,7 @@ class Gateway extends OffsiteGateway
     /**
      * @inheritdoc
      */
-    public function getSettingsHtml()
+    public function getSettingsHtml(): ?string
     {
         return Craft::$app->getView()->renderTemplate('commerce-mollie-plus/gatewaySettings', ['gateway' => $this]);
     }
@@ -150,15 +190,15 @@ class Gateway extends OffsiteGateway
     /**
      * @inheritdoc
      */
-    public function getTransactionHashFromWebhook()
+    public function getTransactionHashFromWebhook(): ?string
     {
         return Craft::$app->getRequest()->getParam('commerceTransactionHash');
     }
-    
+
     /**
      * @inheritdoc
      */
-    public function getPaymentFormHtml(array $params)
+    public function getPaymentFormHtml(array $params): ?string
     {
         try {
             $defaults = [
@@ -167,7 +207,7 @@ class Gateway extends OffsiteGateway
                 'paymentMethods' => $this->fetchPaymentMethods(['resource' => 'orders']),
                 'issuers' => $this->fetchIssuers(),
             ];
-        } catch (\Throwable $exception) {
+        } catch (\Throwable) {
             // In case this is not allowed for the account
             return parent::getPaymentFormHtml($params);
         }
@@ -187,22 +227,32 @@ class Gateway extends OffsiteGateway
 
     /**
      * @param array $parameters
-     * @return mixed
+     * @return array
+     * @throws InvalidRequestException
      */
-    public function fetchPaymentMethods(array $parameters = [])
+    public function fetchPaymentMethods(array $parameters = []): array
     {
-        $paymentMethodsRequest = $this->createGateway()->fetchPaymentMethods($parameters);
+        /** @var OmnipayGateway $gateway */
+        $gateway = $this->createGateway();
 
-        return $paymentMethodsRequest->sendData($paymentMethodsRequest->getData())->getPaymentMethods();
+        $paymentMethodsRequest = $gateway->fetchPaymentMethods($parameters);
+
+        /** @var FetchPaymentMethodsResponse $response */
+        $response = $paymentMethodsRequest->sendData($paymentMethodsRequest->getData());
+
+        return $response->getPaymentMethods();
     }
 
     /**
      * @param array $parameters
-     * @return mixed
+     * @return array
+     * @throws InvalidRequestException
      */
-    public function fetchIssuers(array $parameters = [])
+    public function fetchIssuers(array $parameters = []): array
     {
-        $issuersRequest = $this->createGateway()->fetchIssuers($parameters);
+        /** @var OmnipayGateway $gateway */
+        $gateway = $this->createGateway();
+        $issuersRequest = $gateway->fetchIssuers($parameters);
 
         return $issuersRequest->sendData($issuersRequest->getData())->getIssuers();
     }
@@ -210,9 +260,9 @@ class Gateway extends OffsiteGateway
     /**
      * @inheritdoc
      */
-    public function populateRequest(array &$request, BasePaymentForm $paymentForm = null)
+    public function populateRequest(array &$request, BasePaymentForm $paymentForm = null): void
     {
-        if ($paymentForm) {
+        if ($paymentForm !== null) {
             /** @var MollieOffsitePaymentForm $paymentForm */
             if ($paymentForm->paymentMethod) {
                 $request['paymentMethod'] = $paymentForm->paymentMethod;
@@ -231,10 +281,29 @@ class Gateway extends OffsiteGateway
     {
         /** @var OmnipayGateway $mollieGateway */
         $mollieGateway = $this->gateway();
-        
-        $request = $mollieGateway->createOrder($request);
 
-        return $request;
+        /** @var ItemBag $orderItems */
+        $orderItems = $request['items'];
+
+        $items = [];
+        /** @var Item $orderItem */
+        foreach ($orderItems->all() as $orderItem) {
+            $items[] = [
+                'type' => $orderItem->getType(),
+                'name' => $orderItem->getName(),
+                'sku' => $orderItem->getSku(),
+                'quantity' => $orderItem->getQuantity(),
+                'unitPrice' => $orderItem->getUnitPrice(),
+                'discountAmount' => $orderItem->getDiscountAmount(),
+                'vatRate' => $orderItem->getVatRate(),
+                'vatAmount' => $orderItem->getVatAmount(),
+                'totalAmount' => $orderItem->getTotalAmount(),
+            ];
+        }
+
+        $request['items'] = $items;
+
+        return $mollieGateway->createOrder($request);
     }
 
     /**
@@ -242,14 +311,15 @@ class Gateway extends OffsiteGateway
      */
     protected function prepareRefundRequest($request, string $reference): RequestInterface
     {
-        $res = $this->gateway()
-            ->fetchOrder(['transactionReference' => $reference, 'includePayments' => true])
+        /** @var OmnipayGateway $gateway */
+        $gateway = $this->gateway();
+        $res = $gateway->fetchOrder(['transactionReference' => $reference, 'includePayments' => true])
             ->send();
 
         if (!$res->isSuccessful()) {
             throw new \Exception(Craft::t('commerce-mollie-plus', 'Mollie Order #{reference} not found.', ['reference' => $reference]));
         }
-        
+
         // Find a payment suitable for refund
         $data = $res->getData();
         $payments = $data['_embedded']['payments'] ?? [];
@@ -259,13 +329,12 @@ class Gateway extends OffsiteGateway
                 $targetPaymentId = $payment['id'];
             }
         }
-        
+
         if ($targetPaymentId === null) {
             throw new \Exception(Craft::t('commerce-mollie-plus', 'Unable to find a payment to refund.'));
         }
-        
-        /** @var AbstractRequest $refundRequest */
-        $refundRequest = $this->gateway()->refund($request);
+
+        $refundRequest = $gateway->refund($request);
         $refundRequest->setTransactionReference($targetPaymentId);
 
         return $refundRequest;
@@ -295,9 +364,7 @@ class Gateway extends OffsiteGateway
         /** @var OmnipayGateway $mollieGateway */
         $mollieGateway = $this->gateway();
 
-        $request = $mollieGateway->completeOrder($request);
-
-        return $request;
+        return $mollieGateway->completeOrder($request);
     }
 
     /**
@@ -309,9 +376,8 @@ class Gateway extends OffsiteGateway
         $mollieGateway = $this->gateway();
 
         $request['transactionReference'] = $reference;
-        $request = $mollieGateway->createShipment($request);
 
-        return $request;
+        return $mollieGateway->createShipment($request);
     }
 
 
@@ -327,13 +393,14 @@ class Gateway extends OffsiteGateway
         $transactionHash = $this->getTransactionHashFromWebhook();
         $transaction = Commerce::getInstance()->getTransactions()->getTransactionByHash($transactionHash);
 
-        if (!$transaction) {
-            Craft::warning('Transaction with the hash “'.$transactionHash.'“ not found.', 'commerce');
+        if (!$transaction instanceof \craft\commerce\models\Transaction) {
+            Craft::warning('Transaction with the hash “' . $transactionHash . '“ not found.', 'commerce');
             $response->data = 'ok';
 
             return $response;
         }
 
+        /** @var OmnipayGateway $gateway */
         $gateway = $this->createGateway();
 
         $request = $gateway->fetchOrder(['transactionReference' => $transaction->reference]);
@@ -350,7 +417,7 @@ class Gateway extends OffsiteGateway
         $childTransaction->type = $transaction->type;
 
         Craft::info('Mollie ORDER STATUS received: ' . $res->getStatus(), 'commerce-mollie-plus');
-        
+
         if ($res->isPaid() || $res->getStatus() === 'completed') {
             // Try to find child successful authorize transaction and if found, make it the parent one
             if ($transaction->status != TransactionRecord::STATUS_SUCCESS) {
@@ -359,54 +426,49 @@ class Gateway extends OffsiteGateway
                     'status' => TransactionRecord::STATUS_SUCCESS,
                     'type' => TransactionRecord::TYPE_AUTHORIZE,
                 ])->limit(1)->select(['id'])->scalar();
-                
+
                 if ($authorizeTransactionId) {
                     $authorizeTransaction = Commerce::getInstance()->getTransactions()->getTransactionById($authorizeTransactionId);
-                    if ($authorizeTransaction) {
+                    if ($authorizeTransaction !== null) {
                         $transaction = $authorizeTransaction;
                         $childTransaction = Commerce::getInstance()->getTransactions()->createTransaction(null, $authorizeTransaction);
                         $childTransaction->type = $transaction->type;
                     }
                 }
             }
-            
             $successfulCaptureChildTransaction = TransactionRecord::find()->where([
                 'parentId' => $transaction->id,
                 'status' => TransactionRecord::STATUS_SUCCESS,
                 'type' => TransactionRecord::TYPE_CAPTURE,
             ])->count();
-
             if ($successfulCaptureChildTransaction) {
-                Craft::warning('Successful capture child transaction for “'.$transactionHash.'“ already exists.', 'commerce');
+                Craft::warning('Successful capture child transaction for “' . $transactionHash . '“ already exists.', 'commerce');
                 $response->data = 'ok';
 
                 return $response;
             }
-            
             $childTransaction->type = TransactionRecord::TYPE_CAPTURE;
             $childTransaction->status = TransactionRecord::STATUS_SUCCESS;
-        } else if ($res->isAuthorized()) {
+        } elseif ($res->isAuthorized()) {
             // Check to see if a successful authorize child transaction already exist and skip out early if they do
             $successfulAuthorizeChildTransaction = TransactionRecord::find()->where([
                 'parentId' => $transaction->id,
                 'status' => TransactionRecord::STATUS_SUCCESS,
                 'type' => TransactionRecord::TYPE_AUTHORIZE,
             ])->count();
-
             if ($successfulAuthorizeChildTransaction) {
-                Craft::warning('Successful authorize child transaction for “'.$transactionHash.'“ already exists.', 'commerce');
+                Craft::warning('Successful authorize child transaction for “' . $transactionHash . '“ already exists.', 'commerce');
                 $response->data = 'ok';
 
                 return $response;
             }
-
             $childTransaction->type = TransactionRecord::TYPE_AUTHORIZE;
             $childTransaction->status = TransactionRecord::STATUS_SUCCESS;
-        } else if ($res->isExpired()) {
+        } elseif ($res->isExpired()) {
             $childTransaction->status = TransactionRecord::STATUS_FAILED;
-        } else if ($res->isCancelled()) {
+        } elseif ($res->isCancelled()) {
             $childTransaction->status = TransactionRecord::STATUS_FAILED;
-        } else if (isset($this->data['status']) && 'failed' === $this->data['status']) {
+        } elseif (isset($this->data['status']) && 'failed' === $this->data['status']) {
             $childTransaction->status = TransactionRecord::STATUS_FAILED;
         } else {
             $response->data = 'ok';
@@ -422,12 +484,15 @@ class Gateway extends OffsiteGateway
         $response->data = 'ok';
 
 
-        Craft::info('Transaction created: #'.$childTransaction->id, 'commerce-mollie-plus');
+        Craft::info('Transaction created: #' . $childTransaction->id, 'commerce-mollie-plus');
 
         return $response;
     }
 
-    public function getOrderStatusOptions($optional = null)
+    /**
+     * @return array<int, array{value: string|null, label: mixed}>
+     */
+    public function getOrderStatusOptions($optional = null): array
     {
         $statuses = Commerce::getInstance()->getOrderStatuses()->getAllOrderStatuses();
         $options = [];
@@ -437,40 +502,39 @@ class Gateway extends OffsiteGateway
         }
 
         foreach ($statuses as $status) {
-            $options[] = ['value' => $status->handle, 'label' => $status->displayName];
+            $options[] = ['value' => $status->handle, 'label' => $status->getDisplayName()];
         }
 
         return $options;
     }
-    
+
     /**
      * @inheritdoc
      */
-    protected function createPaymentRequest(Transaction $transaction, $card = null, $itemBag = null): array
+    protected function createPaymentRequest(Transaction $transaction, ?CreditCard $card = null, ?ItemBag $itemBag = null): array
     {
-        if ($card !== null) {
-            $card->setPhone(null);
-        }
-        
+        $card?->setPhone('');
+
         $request = parent::createPaymentRequest($transaction, $card, $itemBag);
-        $request['orderNumber'] = $transaction->order->number;
-        
+        $request['orderNumber'] = $transaction->getOrder()->number;
+
         if (!empty($transaction->note)) {
             $request['description'] = $transaction->note;
         }
 
-        $orderLanguage = $transaction->order->orderLanguage;
+        $orderLanguage = $transaction->getOrder()->orderLanguage;
         $orderLanguage = str_replace('-', '_', $orderLanguage);
 
         // If the Craft locale isn't language specific, add it
         if (!str_contains($orderLanguage, '_')) {
-            $orderLanguage = $orderLanguage . '_' . strtoupper($orderLanguage);
+            $orderLanguage .= '_' . strtoupper($orderLanguage);
         }
 
         // Check if the $orderLanguage is in the list of Mollie languages
         // Otherwise fallback to en_US
         if (!in_array($orderLanguage, [
             'en_US',
+            'en_GB',
             'nl_NL',
             'nl_BE',
             'fr_FR',
@@ -490,21 +554,20 @@ class Gateway extends OffsiteGateway
             'hu_HU',
             'pl_PL',
             'lv_LV',
-            'lt_LT'
+            'lt_LT',
         ])) {
             $orderLanguage = 'en_US';
         }
 
         $request['locale'] = $orderLanguage;
 
-
-        if($this->hasEventHandlers(static::EVENT_CREATE_PAYMENT_REQUEST)) {
+        if ($this->hasEventHandlers(static::EVENT_CREATE_PAYMENT_REQUEST)) {
             $event = new CreatePaymentRequestEvent(['request' => $request, 'transaction' => $transaction]);
             $this->trigger(static::EVENT_CREATE_PAYMENT_REQUEST, $event);
 
             $request = $event->request;
         }
-        
+
         return $request;
     }
 
@@ -519,45 +582,65 @@ class Gateway extends OffsiteGateway
         $count = -1;
 
         foreach ($order->getLineItems() as $item) {
-            $price = Currency::round($item->salePrice);
-            
-            if ($price !== 0) {
-                $count++;
-                /** @var Purchasable $purchasable */
-                $purchasable = $item->getPurchasable();
-                $defaultDescription = Craft::t('commerce', 'Item ID').' '.$item->id;
-                $purchasableDescription = $purchasable ? $purchasable->getDescription() : $defaultDescription;
-                $description = isset($item->snapshot['description']) ? $item->snapshot['description'] : $purchasableDescription;
-                $description = empty($description) ? 'Item '.$count : $description;
+            $price = Currency::round($item->getSalePrice());
 
-                $totalTax = $item->getTaxIncluded();
-                
+            if ($price != 0) {
+                $count++;
+                $defaultDescription = Craft::t('commerce', 'Item ID') . ' ' . $item->id;
+                $description = !empty($item->getDescription()) ? $item->getDescription() : $defaultDescription;
+
                 $vatRate = null;
+                $taxIncluded = false;
+                $itemShippingRate = 0;
                 foreach ($item->getAdjustments() as $adjustment) {
-                    if ($adjustment->type == 'tax' && $adjustment->included) {
+                    if ($adjustment->type == 'tax') {
+                        if ($adjustment->included) {
+                            $taxIncluded = true;
+                        }
                         $snapshot = $adjustment->getSourceSnapshot();
                         if (isset($snapshot['rate'])) {
                             $vatRate = $snapshot['rate'];
-                            break;
+                            continue;
                         }
                     }
+                    if ($adjustment->type == 'shipping') {
+                        $snapshot = $adjustment->getSourceSnapshot();
+                        if (isset($snapshot['perItemRate'])) {
+                            $itemShippingRate = $snapshot['perItemRate'];
+                        }
+                    }
+                }
+
+                if ($taxIncluded) {
+                    $totalTax = $item->getTaxIncluded();
+                } else {
+                    $totalTax = $item->getTax();
                 }
 
                 if ($vatRate === null) {
                     $vatRate = ($totalTax) / (($price * $item->qty) - $totalTax);
                 }
-                
-                $items[] = [
-                    'name' => $description,
-                    'description' => $description,
-                    'quantity' => $item->qty,
-                    'price' => $price,
-                    'unitPrice' => $price,
-                    'vatRate' => sprintf('%0.2f', $vatRate * 100),
-                    'vatAmount' => $totalTax,
-                ];
 
-                $priceCheck += ($item->qty * $item->salePrice);
+                $totalPrice = $item->getTotal();
+
+                $vatAmountUnit = 0;
+                if (!$taxIncluded) {
+                    $vatAmountUnit = Currency::round($totalTax / $item->qty);
+                }
+
+                $items[] = $this->setMollieLineItem(
+                    'physical',
+                    $description,
+                    $item->qty,
+                    $price + $vatAmountUnit + $itemShippingRate,
+                    abs($item->getDiscount()),
+                    sprintf('%0.2f', $vatRate * 100),
+                    $totalTax,
+                    $totalPrice,
+                    $item->getSku()
+                );
+
+                $priceCheck += $totalPrice;
             }
         }
 
@@ -565,23 +648,59 @@ class Gateway extends OffsiteGateway
 
         foreach ($order->getAdjustments() as $adjustment) {
             $price = Currency::round($adjustment->amount);
-            if (($adjustment->included == 0 || $adjustment->included == false) && $price !== 0) {
+            $name = empty($adjustment->name) ? $adjustment->type . " " . $count : $adjustment->name . (!empty($adjustment->description) ? " - " . $adjustment->description : '');
+            if ($adjustment->type == 'shipping' && !$adjustment->included && !$adjustment->lineItemId && $price !== 0.0) {
                 $count++;
-                $items[] = [
-                    'name' => empty($adjustment->name) ? $adjustment->type." ".$count : $adjustment->name,
-                    'description' => empty($adjustment->description) ? $adjustment->type.' '.$count : $adjustment->description,
-                    'quantity' => 1,
-                    'price' => $price,
-                    'unitPrice' => $price,
-                    'vatRate' => '0.00',
-                ];
+                $items[] = $this->setMollieLineItem(
+                    'shipping_fee',
+                    $name,
+                    1,
+                    $price,
+                    0.0,
+                    '0.00',
+                    0.0,
+                    $price
+                );
+
+                $priceCheck += $adjustment->amount;
+            }
+
+            if ($adjustment->type == 'discount' && !$adjustment->included && !$adjustment->lineItemId && $price !== 0.0) {
+                $count++;
+                $items[] = $this->setMollieLineItem(
+                    'discount',
+                    $name,
+                    1,
+                    $price,
+                    0.0,
+                    '0.00',
+                    0.0,
+                    $price
+                );
+
+                $priceCheck += $adjustment->amount;
+            }
+
+            if ($adjustment->type == 'tax' && !$adjustment->included && !$adjustment->lineItemId && $price !== 0.0) {
+                $count++;
+                $items[] = $this->setMollieLineItem(
+                    'physical',
+                    $name,
+                    1,
+                    $price,
+                    0.0,
+                    '0.00',
+                    0.0,
+                    $price
+                );
+
                 $priceCheck += $adjustment->amount;
             }
         }
 
         $priceCheck = Currency::round($priceCheck);
-        $totalPrice = Currency::round($order->totalPrice);
-        $same = (bool)($priceCheck === $totalPrice);
+        $totalPrice = Currency::round($order->getTotalPrice());
+        $same = $priceCheck === $totalPrice;
 
         if (!$same) {
             Craft::error('Item bag total price does not equal the orders totalPrice, some payment gateways will complain.', __METHOD__);
@@ -593,9 +712,12 @@ class Gateway extends OffsiteGateway
     /**
      * @inheritdoc
      */
-    protected function createItemBagForOrder(Order $order)
+    protected function createItemBagForOrder(Order $order): ?ItemBag
     {
-        return $this->getItemListForOrder($order);
+        $items = $this->getItemListForOrder($order);
+        $itemBagClassName = $this->getItemBagClassName();
+
+        return new $itemBagClassName($items);
     }
 
     /**
@@ -606,7 +728,7 @@ class Gateway extends OffsiteGateway
         /** @var OmnipayGateway $gateway */
         $gateway = static::createOmnipayGateway($this->getGatewayClassName());
 
-        $gateway->setApiKey(Craft::parseEnv($this->apiKey));
+        $gateway->setApiKey($this->getApiKey());
 
         $plugin = Craft::$app->getPlugins()->getPluginInfo('commerce-mollie-plus');
         if ($plugin) {
@@ -617,7 +739,7 @@ class Gateway extends OffsiteGateway
         if ($commerce) {
             $gateway->addVersionString('CraftCommerce/' . $commerce['version']);
         }
-        
+
         $gateway->addVersionString('Craft/' . Craft::$app->getVersion());
         $gateway->addVersionString('uap/eSEEG6szENBK5BjD');
 
@@ -627,7 +749,7 @@ class Gateway extends OffsiteGateway
     /**
      * @inheritdoc
      */
-    protected function getGatewayClassName()
+    protected function getGatewayClassName(): string
     {
         return '\\' . OmnipayGateway::class;
     }
@@ -637,6 +759,35 @@ class Gateway extends OffsiteGateway
      */
     protected function prepareResponse(ResponseInterface $response, Transaction $transaction): RequestResponseInterface
     {
+        /** @var AbstractResponse $response */
         return new RequestResponse($response, $transaction);
+    }
+
+    /**
+     * @param string $type
+     * @param string $name
+     * @param int $quantity
+     * @param float $unitPrice
+     * @param float $discountAmount
+     * @param string $vatRate
+     * @param float $vatAmount
+     * @param float $totalAmount
+     * @param string $sku
+     * @return Item
+     */
+    private function setMollieLineItem(string $type, string $name, int $quantity, float $unitPrice, float $discountAmount, string $vatRate, float $vatAmount, float $totalAmount, string $sku = ''): Item
+    {
+        $mollieItem = new Item();
+        $mollieItem->setType($type);
+        $mollieItem->setName($name);
+        $mollieItem->setSku($sku);
+        $mollieItem->setQuantity($quantity);
+        $mollieItem->setUnitPrice($unitPrice);
+        $mollieItem->setDiscountAmount($discountAmount);
+        $mollieItem->setVatRate($vatRate);
+        $mollieItem->setVatAmount($vatAmount);
+        $mollieItem->setTotalAmount($totalAmount);
+
+        return $mollieItem;
     }
 }