+++ src/models/forms/MollieOffsitePaymentForm.php
@@ -2,7 +2,6 @@
 
 namespace white\commerce\mollie\plus\models\forms;
 
-use Craft;
 use craft\commerce\models\payments\BasePaymentForm;
 
 class MollieOffsitePaymentForm extends BasePaymentForm